pyproject.toml
[project]
name = "biojson"
version = "0.1.0"
description = "BioJSON - 植物营养代谢基因结构化提取流水线 + RAG 问答系统"
requires-python = ">=3.11"

dependencies = [
    # ---- Web 框架 ----
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.7",
    "flask>=3.0.0",               # admin Blueprint + WSGIMiddleware
    # ---- 安全 & 中间件 ----
    "slowapi>=0.1.9",
    "cachetools>=5.3.0",
    # ---- 认证 & 数据库 ----
    "supabase>=2.0.0",
    # ---- AI / LLM ----
    "openai>=1.0.0",
    # ---- 检索 & 向量 ----
    "numpy>=1.24.0",
    "requests>=2.31.0",
    # ---- 生物信息学 (CRISPR pipeline) ----
    "lxml>=5.0.0",
    "biopython>=1.80",
    # ---- 配置 ----
    "python-dotenv>=1.0.0",
    "httpx[socks]>=0.28.1",
    "claude-code-sdk>=0.0.25",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
# 个人知识库 PDF 解析
pdf = ["PyMuPDF>=1.23.0"]
# TF-IDF 索引构建
index = ["scikit-learn>=1.3.0"]
# 完整开发/测试环境
dev = [
    "PyMuPDF>=1.23.0",
    "scikit-learn>=1.3.0",
    "pytest>=7.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
admin = [
    "static/*",
]
server = [
    "static/*",
    "static/**/*",
]
skills = [
    "*/skill.md",
    "*/SKILL.md",
]
crispr = [
    "templates/*.txt",
]
indexing = [
    "resources/*.json",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests", "extractor/tests"]
markers = [
    "integration: tests that call real external services and require configured API keys",
    "e2e: browser end-to-end tests that exercise full user flows",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
